add DisLoRATrainer
